Move all OAuth code to Shoebox
