+++ kifi-backend/modules/common/app/com/keepit/common/oauth/RichIdentity.scala
@@ -3,63 +3,14 @@ package com.keepit.common.oauth
 import com.keepit.common.mail.EmailAddress
 import com.keepit.slack.models._
 import com.keepit.social.IdentityHelpers
-import securesocial.core.providers.utils.GravatarHelper
 import securesocial.core._
-import com.keepit.common.core._
+import securesocial.core.providers.utils.GravatarHelper
 
 sealed trait RichIdentity
 
 case class FacebookIdentity(socialUser: SocialUser) extends RichIdentity
-object FacebookIdentity {
-  def apply(auth: OAuth2Info, info: UserProfileInfo): FacebookIdentity = {
-    val socialUser = SocialUser(
-      identityId = IdentityId(info.userId.id, ProviderIds.Facebook.id),
-      firstName = info.firstNameOpt getOrElse "",
-      lastName = info.lastNameOpt getOrElse "",
-      fullName = info.name,
-      avatarUrl = info.pictureUrl.map(_.toString),
-      email = info.emailOpt.map(_.address),
-      authMethod = AuthenticationMethod.OAuth2,
-      oAuth2Info = Some(auth)
-    )
-    FacebookIdentity(socialUser)
-  }
-}
-
 case class LinkedInIdentity(socialUser: SocialUser) extends RichIdentity
-object LinkedInIdentity {
-  def apply(auth: OAuth2Info, info: UserProfileInfo): LinkedInIdentity = {
-    val socialUser = SocialUser(
-      identityId = IdentityId(info.userId.id, ProviderIds.LinkedIn.id),
-      firstName = info.firstNameOpt getOrElse "",
-      lastName = info.lastNameOpt getOrElse "",
-      fullName = info.name,
-      avatarUrl = info.pictureUrl.map(_.toString),
-      email = info.emailOpt.map(_.address),
-      authMethod = AuthenticationMethod.OAuth2,
-      oAuth2Info = Some(auth)
-    )
-    LinkedInIdentity(socialUser)
-  }
-}
-
 case class TwitterIdentity(socialUser: SocialUser, pictureUrl: Option[String], profileUrl: Option[String]) extends RichIdentity
-object TwitterIdentity {
-  def apply(auth: OAuth1Info, info: TwitterUserInfo): TwitterIdentity = {
-    val socialUser = SocialUser(
-      identityId = IdentityId(info.id.toString, ProviderIds.Twitter.id),
-      firstName = info.firstName,
-      lastName = info.lastName,
-      fullName = info.name,
-      avatarUrl = info.pictureUrl,
-      email = None,
-      authMethod = AuthenticationMethod.OAuth1,
-      oAuth1Info = Some(auth)
-    )
-    TwitterIdentity(socialUser, info.pictureUrl, Some(info.profileUrl))
-  }
-}
-
 case class EmailPasswordIdentity(firstName: String, lastName: String, email: EmailAddress, password: Option[PasswordInfo]) extends RichIdentity
 
 case class SlackIdentity(teamId: SlackTeamId, teamName: SlackTeamName, token: Option[SlackUserAccessToken], scopes: Set[SlackAuthScope], userId: SlackUserId, username: SlackUsername, user: Option[SlackUserInfo]) extends RichIdentity
