+++ kifi-backend/modules/shoebox/app/com/keepit/common/oauth/OAuth2Configuration.scala
