+++ kifi-backend/modules/shoebox/app/com/keepit/common/oauth/adaptor/SecureSocialOAuth2Adaptor.scala
