+++ kifi-backend/modules/shoebox/test/com/keepit/social/FakeTwitterSocialGraph.scala
@@ -47,7 +47,7 @@ class FakeTwitterSocialGraph @Inject() (
   val twtrOAuthProvider = new TwitterOAuthProviderImpl(airbrake, oauth1Config) {
     override def getRichIdentity(accessToken: OAuth1TokenInfo): Future[TwitterIdentity] = Future.successful {
       val info = tweetfortytwoInfo.copy(screenName = "tweet42")
-      TwitterIdentity(accessToken, info)
+      TwitterOAuthProvider.toIdentity(accessToken, info)
     }
   }
